Refactoring
- Moved model utility class.
