+++ app/Models/ItemTransfer.php
@@ -3,8 +3,6 @@ declare(strict_types=1);
 
 namespace App\Models;
 
-use App\Utilities\General;
-use App\Utilities\Model as ModelUtility;
 use Illuminate\Database\Eloquent\Builder as QueryBuilder;
 use Illuminate\Database\Eloquent\Model;
 
@@ -65,7 +63,7 @@ class ItemTransfer extends Model
             $collection->where($this->table .'.item_id', '=', $parameters['item']);
         }
 
-        $collection = ModelUtility::applyResourceTypeCollectionCondition(
+        $collection = Clause::applyResourceTypeCollectionCondition(
             $collection,
             $permitted_resource_types,
             $include_public
@@ -145,7 +143,7 @@ class ItemTransfer extends Model
             $collection->where($this->table .'.item_id', '=', $parameters['item']);
         }
 
-        $collection = ModelUtility::applyResourceTypeCollectionCondition(
+        $collection = Clause::applyResourceTypeCollectionCondition(
             $collection,
             $permitted_resource_types,
             $include_public
