+++ app/Models/ResourceTypeItem/SimpleExpense.php
@@ -5,7 +5,7 @@ namespace App\Models\ResourceTypeItem;
 
 use App\Interfaces\ResourceTypeItem\IModel;
 use App\Request\Validate\Boolean;
-use App\Utilities\Model as ModelUtility;
+use App\Models\Clause;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Query\Builder as QueryBuilder;
 
@@ -62,13 +62,13 @@ class SimpleExpense extends Model implements IModel
             }
         }
 
-        $collection = ModelUtility::applySearch(
+        $collection = Clause::applySearch(
             $collection,
             $this->item_table,
             $search_parameters
         );
 
-        $collection = ModelUtility::applyFiltering(
+        $collection = Clause::applyFiltering(
             $collection,
             $this->item_table,
             $filter_parameters
@@ -179,13 +179,13 @@ class SimpleExpense extends Model implements IModel
             }
         }
 
-        $collection = ModelUtility::applySearch(
+        $collection = Clause::applySearch(
             $collection,
             $this->item_table,
             $search_parameters
         );
 
-        $collection = ModelUtility::applyFiltering(
+        $collection = Clause::applyFiltering(
             $collection,
             $this->item_table,
             $filter_parameters
