+++ app/Models/ResourceTypeItem/Summary/AllocatedExpense.php
@@ -5,7 +5,7 @@ namespace App\Models\ResourceTypeItem\Summary;
 
 use App\Interfaces\ResourceTypeItem\ISummaryModelCategories;
 use App\Interfaces\ResourceTypeItem\ISummaryModel;
-use App\Utilities\Model as ModelUtility;
+use App\Models\Clause;
 use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Query\Builder as QueryBuilder;
@@ -359,13 +359,13 @@ class AllocatedExpense extends Model implements ISummaryModel, ISummaryModelCate
             $collection->whereRaw(DB::raw("MONTH({$this->sub_table}.effective_date) = {$month}"));
         }
 
-        $collection = ModelUtility::applySearch(
+        $collection = Clause::applySearch(
             $collection,
             $this->sub_table,
             $search_parameters
         );
 
-        $collection = ModelUtility::applyFiltering(
+        $collection = Clause::applyFiltering(
             $collection,
             $this->sub_table,
             $filter_parameters
